requirements.txt
Django==2.2
django-cors-headers==2.5.3
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0